Fix error with hide values.
When using flags `--geo.type` and `metric.hideip` simultaneously, ip wasn't sent to getIPDetails* functions. Now it fixed.
A wrapper has been written that hides the values of the user and the IP in the metric, if the `--metric.hideip`  and  `--metric.hideuser` flags are used.
Also use local variables inside parseLine func.